+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Note that since we don't clearly distinguish between a public and private interf
 - Fix visual for bonds between two aromatic rings
 - Fix visual for delocalized bonds (parsed from mmcif and mol2)
 - Fix ring computation algorithm
+- Add ``UnitResonance`` property with info about delocalized triplets
 
 ## [v3.2.0] - 2022-02-17